Treat the access token as expired a few minutes before its actual expiry, #16
When sending a request a few moments before the token's expiry time, it is
possible the server denies it with a "401: Access token expired". This could
be caused by e.g. latency or server time mismatches. We resolve it by treating
the access token as expired a few minutes before it actually is.